simpleperf: binary_cache_builder: add option to get every binary
Right now binary_cache_builder.py considers only files that are
actually mentioned in samples. This is an issue for perf.data files
that are not sample based, like those containing ETM traces.

This change exposes the contents of the BUILD_ID feature secton to
Python through libsimpleperf_report.so.

A new option, "--every" is added to binary_cache_builder.py to make
use of this to download every file recorded with a build id from the
device.

Test: ./test.py -p 'TestReportLib*'
Test: ./test.py -p 'TestBinaryCacheBuilder*'

Change-Id: I15ed80bf997cbbf769a5f979b23dfdbc7ad5f766