simpleperf: inject: Add option to search binaries with filename
When --symdir is given, binaries are searched by build id.
This patch adds --allow-mismatched-build-id in the inject
cmd. When used, it searches binaries by filename and allows
missing or mismatched build ids. It also generates a message
to show the found binary. So users can check if that is correct.

Bug: 361638495
Test: run simpleperf_unit_test
Test: run simpleperf inject --allow-mismatched-build-id
Change-Id: I0c34eeca2b6cd6ec860d7ea97935b97933969b97