Add LoginLauncherActivity in LauncherStart event
HSUM tablets and desktop have the option of using a logout flow - and a
corresponding Login screen. This screen is drawn by the Login app -
which is the launcher of HSU0 (headless system user 0).

Performance tests on HSUM devices need to search for the default
activity of Login app - LoginLauncherActivity - to understand that an
HSUM device has booted.

Add LoginLauncherActivity in the LauncherStart event.

Bug: 375051611
Flag: EXEMPT desktop-only change
Test: ./bootanalyze.py -c ./config.yaml -G 4M -t -y -n 10
Change-Id: I4fe9d8dd78acd1ede7cb266f795351ece0e287f4