Put builtin event types into a constexpr table to avoid build time bloat
The std::set initialized here contains EventType, whose initialization
involves constructing a std::string. As a result, the generated static
initializer contains a call to std::string constructor for every string
literal in it. This leads to very long build times (>5min), and bloats
the binary due to a large number of generated function invocations.

A simple solution is to build a constexpr array of what we will use to
initialize the EventType items, and add a helper function that does all
the actual initialization.

The initialization cost should be around the same. Instead of generating
one std::string constructor call per literal, we are now simply calling
that in a loop against the literal table. The std::set construction is
done by insert() instead of batch insert in constructor, which is
slightly more expensive but not significantly (the complexity remains N
log N).

With the change, building event_table takes 5 seconds and the binary
size is down to 220KB from 392KB.

Bug: 381802323
Test: arm64-v8a simpleperf_unit_test:simpleperf_unit_test
Change-Id: Ibece7804895c55c1688cb4c13058656c63016957