simpleperf: Prefer running simpleperf through simpleperf_app_runner
When `-app` is used to record profiles for an app, simpleperf
uses either run-as or simpleperf_app_runner to switch to the
app's domain.
Since run-as will lose the ability of recording kernel events,
it's better to prefer simpleperf_app_runner.
The run-as support needs to be kept to support Android P
for studio profiler.

Bug: 390626125
Test: run CtsSimpleperfTestCases
Change-Id: I96b4d6a634f2816f811777f81ebaedeb425194cd