simpleperf: update record_cmd.addr_filter_option test.
Using the default output of simpleperf inject, the lines describing
the DSOs currently look like these:

// build_id: 0xfc65a0b5f49361c5774d6b9fe66b076600000000
// /system/bin/toybox

The current code does not detect these lines, however the test does
not fail as it does not check if the sought after DSO was found, only
that no other DSO dares to show itself.

This change makes sure that what is looked for is found, and updates
the string constants to handle the current output format.

Test: simpleperf_unit_test --gtest_filter=*.addr_filter_option
Change-Id: Ia72f9c54910248debce2b999e8350c917190c5ce